requirements.txt
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
dj-database-url==0.5.0
Django==1.11.17
django-ckeditor==5.5.0
django-countries==5.3.1
django-embed-video==1.1.2
django-js-asset==1.1.0
django-modeltranslation==0.13
django-widget-tweaks==1.4.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.11.3
idna==2.7
ipython-genutils==0.2.0
jsonschema==2.6.0
jupyter-core==4.4.0
nbformat==4.4.0
numpy==1.15.1
oauth2client==4.1.3
pandas==0.23.4
Pillow==5.2.0
plotly==2.7.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
requests==2.19.1
retrying==1.3.3
rsa==3.4.2
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.23
